+++ hikka/forbid_joins.py
@@ -54,9 +54,7 @@ def install_join_forbidder(client: TelegramClient) -> TelegramClient:
                         and hasattr(frame_info.frame, "f_locals")
                         and isinstance(frame_info.frame.f_locals, dict)
                         and "self" in frame_info.frame.f_locals
-                        and isinstance(
-                            frame_info.frame.f_locals["self"], loader.Module
-                        )
+                        and isinstance(frame_info.frame.f_locals["self"], loader.Module)
                         and frame_info.frame.f_locals["self"].__class__.__name__
                         not in {
                             "APIRatelimiterMod",